refactor: apply Suspense to mypage (#62)
